requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.12.1
appdirs==1.4.4
asyncio==4.0.0
attrs==25.4.0
blis==0.7.11
catalogue==2.0.10
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
conllu==6.0.0
cryptography==46.0.5
cymem==2.0.13
distro==1.9.0
docstring_parser==0.17.0
dotmap==1.3.30
en_core_sci_lg @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz#sha256=a2bfde10824958d51cd016b4142563373d553e24767d79b1abf8699596e31a5b
en_core_sci_scibert @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_scibert-0.5.4.tar.gz#sha256=be4afaa2710d63206fafa7864001acbfab1225800bbab577f32765435e29b7f2
faiss-cpu==1.13.2
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
gensim==4.4.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
googletrans==4.0.2
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
h5py==3.16.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.11.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpickle==4.1.1
langcodes==3.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
MedCLIP @ git+https://github.com/RyanWangZf/MedCLIP.git@9c3396f20d5d54e4fae241b8cb06ca45848e98c9
modal==1.3.5
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
neo4j==6.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
nmslib-metabrainz==2.1.3
numpy==1.26.4
openai==2.30.0
owlready2==0.50
packaging==26.0
pandas==3.0.1
pillow==12.1.1
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybind11==3.0.2
pycocoevalcap==1.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pysbd==0.3.4
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rouge_score==0.1.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scispacy==0.6.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
spacy==3.7.5
spacy-alignments==0.9.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.4.0
srsly==2.5.2
sympy==1.14.0
synchronicity==0.11.1
textaugment==3.0.0
thinc==8.2.5
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.13.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.24.0
typer==0.24.1
typer-slim==0.24.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
wget==3.2
wrapt==2.1.2
yarl==1.23.0
